requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-community
python-dotenv
pypdf
streamlit
faiss-cpu

